rename frcs.model to model
