+++ src/systems/cable/cable-manual-wt-log.ts
@@ -1,15 +1,15 @@
 // Outputs sheet: Cable Manual WT/Log column
+import { Chipping } from '../methods/chipping';
+import { Loading } from '../methods/loading';
+import { calculateMachineCosts } from '../methods/machinecosts';
+import { MoveInCosts } from '../methods/moveincost';
 import {
   Assumptions,
   FrcsInputs,
   FrcsOutputs,
   IntermediateVariables,
   MachineCosts,
-} from '../frcs.model';
-import { Chipping } from '../methods/chipping';
-import { Loading } from '../methods/loading';
-import { calculateMachineCosts } from '../methods/machinecosts';
-import { MoveInCosts } from '../methods/moveincost';
+} from '../model';
 import { CYCCU } from './methods/cyccu';
 import { CYPCU } from './methods/cypcu';
 import { FellwtChipLogOther } from './methods/fellwtchiplogother';
