+++ src/systems/cable/methods/cypcu.ts
@@ -1,5 +1,5 @@
 // CableYarding sheet: II. Cable Yarding, Partial Cut, Unbunched (CYPCU)
-import { Assumptions, FrcsInputs, IntermediateVariables, MachineCosts } from '../../frcs.model';
+import { Assumptions, FrcsInputs, IntermediateVariables, MachineCosts } from '../../model';
 
 function CYPCU(
   assumption: Assumptions,
