+++ src/systems/ground/methods/fellbunch.ts
@@ -1,5 +1,5 @@
 // FellBunch sheet
-import { FrcsInputs, IntermediateVariables, MachineCosts } from '../../frcs.model';
+import { FrcsInputs, IntermediateVariables, MachineCosts } from '../../model';
 
 function FellBunch(
   input: FrcsInputs,
