+++ src/systems/ground/methods/forwarding.ts
@@ -1,5 +1,5 @@
 // Forwarding sheet
-import { Assumptions, FrcsInputs, IntermediateVariables, MachineCosts } from '../../frcs.model';
+import { Assumptions, FrcsInputs, IntermediateVariables, MachineCosts } from '../../model';
 
 function Forwarding(
   assumption: Assumptions,
