+++ src/systems/ground/methods/skidding.ts
@@ -1,5 +1,5 @@
 // Skidding sheet
-import { FrcsInputs, IntermediateVariables, MachineCosts } from '../../frcs.model';
+import { FrcsInputs, IntermediateVariables, MachineCosts } from '../../model';
 
 function Skidding(
   input: FrcsInputs,
