+++ src/systems/methods/inlimits.ts
@@ -1,5 +1,5 @@
 // Outputs sheet - Limits part
-import { FrcsInputs, Limits } from '../frcs.model';
+import { FrcsInputs, Limits } from '../model';
 
 function InLimits(input: FrcsInputs) {
   const limit: Limits = {
